rename bank transactions shared utils path
Rename the shared bank transactions utility file to shared.ts and update imports to use @utils/bankTransactions/shared.
