+++ src/components/BankTransactions/BankTransactions.tsx
@@ -5,8 +5,7 @@ import { debounce } from 'lodash-es'
 import { DisplayState } from '@internal-types/bankTransactions'
 import type { LayerError } from '@utils/api/errorHandler'
 import { isAnyBankAccountSyncing } from '@utils/bankAccount'
-import { type BankTransactionFilters } from '@utils/bankTransactions'
-import { BankTransactionsDateFilterMode } from '@utils/bankTransactions'
+import { type BankTransactionFilters, BankTransactionsDateFilterMode } from '@utils/bankTransactions/shared'
 import { BREAKPOINTS } from '@utils/screenSizeBreakpoints'
 import { unsafeAssertUnreachable } from '@utils/switch/assertUnreachable'
 import { usePreloadCategories } from '@hooks/api/businesses/[business-id]/categories/useCategories'
