+++ src/hooks/features/bankTransactions/useAugmentedBankTransactions.tsx
@@ -8,7 +8,7 @@ import { DataModel, Direction } from '@internal-types/general'
 import { type TagFilterInput } from '@internal-types/tags'
 import { decodeRulesSuggestion } from '@schemas/bankTransactions/categorizationRules/categorizationRule'
 import { isAnyBankAccountSyncing } from '@utils/bankAccount'
-import { type BankTransactionFilters, filterVisibility, type NumericRangeFilter } from '@utils/bankTransactions'
+import { type BankTransactionFilters, filterVisibility, type NumericRangeFilter } from '@utils/bankTransactions/shared'
 import { useBankTransactions, type UseBankTransactionsOptions } from '@hooks/api/businesses/[business-id]/bank-transactions/useBankTransactions'
 import { useLinkedAccounts } from '@hooks/legacy/useLinkedAccounts'
 import { CategorizationRulesContext } from '@contexts/CategorizationRulesContext/CategorizationRulesContext'
