+++ src/hooks/legacy/useReceipts.ts
@@ -6,7 +6,7 @@ import { type BankTransaction } from '@internal-types/bankTransactions'
 import type { FileMetadata } from '@internal-types/fileUpload'
 import { type Awaitable } from '@internal-types/utility/promises'
 import { get, post, postWithFormData } from '@utils/api/authenticatedHttp'
-import { hasReceipts } from '@utils/bankTransactions'
+import { hasReceipts } from '@utils/bankTransactions/shared'
 import { useAuth } from '@hooks/utils/auth/useAuth'
 import { useIntlFormatter } from '@hooks/utils/i18n/useIntlFormatter'
 import { useEnvironment } from '@providers/Environment/EnvironmentInputProvider'
