+++ src/utils/bankTransactions/shared.ts
@@ -4,7 +4,6 @@ import { type BankTransaction, DisplayState, type Split, type SuggestedMatch } f
 import { type DateRange } from '@internal-types/general'
 import { Direction } from '@internal-types/general'
 import type { TagFilterInput } from '@internal-types/tags'
-import { type BankTransactionTaxOption } from '@schemas/bankTransactions/bankTransaction'
 import type { CategoryUpdate } from '@schemas/bankTransactions/categoryUpdate'
 import { makeTagKeyValueFromTag } from '@schemas/tag'
 import { CategorizedCategories, ReviewCategories } from '@components/BankTransactions/constants'
@@ -78,14 +77,6 @@ export const countTransactionsToReview = ({
 export const hasReceipts = (bankTransaction?: BankTransaction) =>
   bankTransaction?.document_ids && bankTransaction.document_ids.length > 0
 
-export const getBankTransactionTaxOptions = (bankTransaction?: BankTransaction): BankTransactionTaxOption[] => {
-  if (!bankTransaction?.tax_options) {
-    return []
-  }
-
-  return Object.values(bankTransaction.tax_options).flat()
-}
-
 export const isTransferMatch = (bankTransaction?: BankTransaction) => {
   return bankTransaction?.match?.details.type === 'Transfer_Match'
 }
@@ -137,7 +128,6 @@ export const buildCategorizeBankTransactionPayloadForSplit = (splits: Split[]): 
     : ({
       type: 'Split',
       entries: splits.map(split => ({
-        // TODO: enforce upstream in the category combobox that split.category is non-null
         category: split.category!.classification!,
         amount: split.amount,
         taxCode: split.taxCode ?? null,
